requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Image Processing
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

pymupdf==1.23.8

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
python-dateutil==2.8.2

# OCR
paddlepaddle
paddleocr

# Mistral AI
mistralai>=1.0.0

# String Matching
thefuzz==0.20.0
python-levenshtein==0.23.0
